Created Boxfile to make testing easier, resolve #2
Added Boxfile so I could have an environment with postgresql and redis available. There are some issues with the go engine, so it doesnt work 100% automagically, but atleast I dont have to configure redis and postgresql on my local machine
Added routines to all cachers for initializing the data store.
Added InitializeDatabase to the Cacher interface.
Fix some issues with postgresql integration. Took out the prepared statements and put in simple queries.
Fixed issue with redis when there is no password.