Add (real) wildcard support
You can already add wildcard records (`CNAME *.example.tld`) with the current design (though this doesn't seem to be validated anywhere), but it doesn't support actually looking up domains that only exist as wildcards. This commit adds support for wildcard lookups by replacing the first element of the domain name with an asterisk and looking up that variation as well. It retains the existing fake wildcard support as a fallback mode for when a wildcard CNAME isn't found.

And yes, there's a test to confirm it actually works as expected.