Add wildcard lookup support
You can already add wildcard records with the current design (though this doesn't seem to be validated anywhere), but it doesn't support actually looking up domains that only exist as wildcards. This commit adds support for wildcard lookups by replacing the first element of the domain name with an asterisk and looking up that variation as well.

And yes, there's a test to confirm it actually works as expected.